test: reorganize consumer offset tests
